+++ src/Test/Vector2fProductTest.java
@@ -1,4 +1,4 @@
-package testOracle;
+package Test;
 
 import static org.junit.Assert.*;
